+++ coles_vs_woolies/search/__init__.py
@@ -1,0 +1,3 @@
+from .merchant import coles, iga, woolies
+
+available_merchants = [coles, woolies, iga]